Packages: Create new spec-parser package
