--- blocks/api/post-parser.js
@@ -1,1 +1,0 @@
-export * from './post.pegjs';
