--- blocks/api/post-parser.native.js
@@ -1,1 +1,0 @@
-export * from './post-grammar-parser-pre-generated';
