+++ blocks/api/test/parser.js
@@ -16,9 +16,7 @@ import {
 	default as parsePegjs,
 	parseWithAttributeSchema,
 	toBooleanAttributeMatcher,
-	createParse,
 } from '../parser';
-import { parse as grammarParsePreGenerated } from '../post-grammar-parser-pre-generated';
 import {
 	registerBlockType,
 	unregisterBlockType,
@@ -549,15 +547,9 @@ describe( 'block parser', () => {
 		} );
 	} );
 
-	describe( 'parse() of pegjs parser', () => {
+	describe( 'parse() of @wordpress/spec-parser', () => {
 		// run the test cases using the PegJS defined parser
 		testCases( parsePegjs );
-	} );
-
-	describe( 'parse() of pre-generated parser', () => {
-		// run the test cases using the pre-generated parser
-		const parsePreGenerated = createParse( grammarParsePreGenerated );
-		testCases( parsePreGenerated );
 	} );
 
 	// encapsulate the test cases so we can run them multiple time but with a different parse() function
