+++ package.json
@@ -35,6 +35,7 @@
 		"@wordpress/keycodes": "file:packages/keycodes",
 		"@wordpress/plugins": "file:packages/plugins",
 		"@wordpress/shortcode": "file:packages/shortcode",
+		"@wordpress/spec-parser": "file:packages/spec-parser",
 		"@wordpress/url": "file:packages/url",
 		"@wordpress/viewport": "file:packages/viewport",
 		"@wordpress/wordcount": "file:packages/wordcount",
@@ -105,7 +106,6 @@
 		"node-sass": "4.9.2",
 		"path-type": "3.0.0",
 		"pegjs": "0.10.0",
-		"pegjs-loader": "0.5.4",
 		"phpegjs": "1.0.0-beta7",
 		"postcss-color-function": "4.0.1",
 		"postcss-loader": "2.1.3",
@@ -145,7 +145,7 @@
 	"scripts": {
 		"prebuild": "npm run check-engines",
 		"clean:packages": "rimraf ./packages/*/build ./packages/*/build-module ./packages/*/build-style",
-		"prebuild:packages": "npm run clean:packages && cross-env INCLUDE_PACKAGES=babel-plugin-import-jsx-pragma,postcss-themes SKIP_JSX_PRAGMA_TRANSFORM=1 node ./bin/packages/build.js",
+		"prebuild:packages": "npm run clean:packages && lerna run build && cross-env INCLUDE_PACKAGES=babel-plugin-import-jsx-pragma,postcss-themes SKIP_JSX_PRAGMA_TRANSFORM=1 node ./bin/packages/build.js",
 		"build:packages": "cross-env EXCLUDE_PACKAGES=babel-plugin-import-jsx-pragma,postcss-themes node ./bin/packages/build.js",
 		"build": "npm run build:packages && cross-env NODE_ENV=production webpack",
 		"check-engines": "check-node-version --package",
