+++ packages/spec-parser/test/index.js
@@ -1,0 +1,23 @@
+/**
+ * Internal dependencies
+ */
+import { parse } from "../";
+
+describe("spec-parser", () => {
+	test("parse() works properly", () => {
+		const result = parse(
+			"<!-- wp:core/more --><!--more--><!-- /wp:core/more -->"
+		);
+
+		expect(result).toMatchInlineSnapshot(`
+Array [
+  Object {
+    "attrs": null,
+    "blockName": "core/more",
+    "innerBlocks": Array [],
+    "innerHTML": "<!--more-->",
+  },
+]
+`);
+	});
+});
