+++ webpack.config.js
@@ -158,10 +158,6 @@ const config = {
 	module: {
 		rules: [
 			{
-				test: /\.pegjs/,
-				use: 'pegjs-loader',
-			},
-			{
 				test: /\.js$/,
 				exclude: /node_modules/,
 				use: 'babel-loader',
